requirements.txt
pandas 
numpy 
python-dotenv
openai

# chapter 04 - multiagent systems
autogenstudio
pyautogen
agentops
crewai[tools]
python-dotenv
langchain


